// src/types/errors/validatedErrors.d.ts
export interface ValidatedErrors {
  'nombre'?: string[]
  'apellido_p'?: string[]
  'apellido_m'?: string[]
  'correo'?: string[]
  'pass'?: string[]
  'estado_civil'?: string[]
  'ciudad'?: string[]
  'fecha_nacimiento'?: string[]
  'actividad1'?: string[]
  'actividad2'?: string[]
  'actividad3'?: string[]
}